+++ model/Beer.go
@@ -1,0 +1,9 @@
+package model
+
+type Beer struct {
+	ID       int
+	Name     string
+	Type     string
+	Detail   string
+	ImageURL string
+}
